requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
werkzeug
sqlalchemy
psycopg2-binary
python-multipart
jinja2
itsdangerous
sse-starlette
aiofiles
